icu_plurals::provider::rules::runtime

Function test_rule

pub fn test_rule(rule: &Rule<'_>, operands: &PluralOperands) -> bool

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. In particular, the `DataProvider` implementations are only guaranteed to match with this version's `*_unstable` providers. Use with caution.