Module icu_plurals::rules::reference

Reference version of the Plural Rules parser, AST and serializer.

Modules

ast

AST provides a set of Syntax Tree Nodes used to store the output of parse method that is used in test_condition method to evaluate whether a given PluralCategory should be used.

Structs

Lexer

Unicode Plural Rule lexer is an iterator over tokens produced from an input string.

Functions

parse

Unicode Plural Rule parser converts an input string into a Rule AST.

parse_condition

Unicode Plural Rule parser converts an input string into an AST.

serialize

Unicode Plural Rule serializer converts an AST to a [String].

test_condition

Function used to test Condition against PluralOperands to identify the appropriate PluralCategory.