icu_plurals

Struct PluralElements

pub struct PluralElements<T>(/* private fields */);

A bag of values for different plural cases.

Implementations

impl<T> PluralElements<T>

pub fn new(other: T) -> Self

Creates a new PluralElements with the given default value.

pub fn zero(&self) -> &T

The value for PluralCategory::Zero

pub fn one(&self) -> &T

The value for PluralCategory::One

pub fn two(&self) -> &T

The value for PluralCategory::Two

pub fn few(&self) -> &T

The value for PluralCategory::Few

pub fn many(&self) -> &T

The value for PluralCategory::Many

pub fn other(&self) -> &T

The value for PluralCategory::Other

pub fn try_into_other(self) -> Option<T>

If the only variant is other, returns Some(other).

Examples

use icu_plurals::PluralElements;

let mut only_other = PluralElements::new("abc").with_one_value(Some("abc"));
assert_eq!(only_other.try_into_other(), Some("abc"));

let mut multi = PluralElements::new("abc").with_one_value(Some("def"));
assert_eq!(multi.try_into_other(), None);

pub fn explicit_zero(&self) -> Option<&T>

The value used when the PluralOperands are exactly 0.

pub fn explicit_one(&self) -> Option<&T>

The value used when the PluralOperands are exactly 1.

pub fn map<B, F: FnMut(T) -> B>(self, f: F) -> PluralElements<B>

Applies a function f to convert all values to another type.

Examples

use icu_plurals::PluralElements;

let x = PluralElements::new(11).with_one_value(Some(15));
let y = x.map(|i| i * 2);

assert_eq!(*y.other(), 22);
assert_eq!(*y.one(), 30);

pub fn try_map<B, E, F: FnMut(T) -> Result<B, E>>( self, f: F, ) -> Result<PluralElements<B>, E>

Applies a function f to convert all values to another type, propagating a possible error.

pub fn for_each<F: FnMut(&T)>(&self, f: F)

Immutably applies a function f to each value.

pub fn try_for_each<E, F: FnMut(&T) -> Result<(), E>>( &self, f: F, ) -> Result<(), E>

Immutably applies a function f to each value, propagating a possible error.

pub fn for_each_mut<F: FnMut(&mut T)>(&mut self, f: F)

Mutably applies a function f to each value.

Examples

use icu_plurals::PluralElements;

let mut x = PluralElements::new(11).with_one_value(Some(15));
x.for_each_mut(|i| *i *= 2);

assert_eq!(*x.other(), 22);
assert_eq!(*x.one(), 30);

pub fn try_for_each_mut<E, F: FnMut(&mut T) -> Result<(), E>>( &mut self, f: F, ) -> Result<(), E>

Mutably applies a function f to each value, propagating a possible error.

pub fn as_ref(&self) -> PluralElements<&T>

Converts from &PluralElements<T> to PluralElements<&T>.

impl<T: PartialEq> PluralElements<T>

pub fn with_zero_value(self, zero: Option<T>) -> Self

Sets the value for PluralCategory::Zero.

pub fn with_one_value(self, one: Option<T>) -> Self

Sets the value for PluralCategory::One.

pub fn with_two_value(self, two: Option<T>) -> Self

Sets the value for PluralCategory::Two.

pub fn with_few_value(self, few: Option<T>) -> Self

Sets the value for PluralCategory::Few.

pub fn with_many_value(self, many: Option<T>) -> Self

Sets the value for PluralCategory::Many.

pub fn with_explicit_zero_value(self, explicit_zero: Option<T>) -> Self

Sets the value for explicit 0.

pub fn with_explicit_one_value(self, explicit_one: Option<T>) -> Self

Sets the value for explicit 1.