Struct icu_plurals::PluralRules

pub struct PluralRules { /* private fields */ }

A struct which provides an ability to retrieve an appropriate Plural Category for a given number.

Examples

use icu::locid::locale;
use icu::plurals::{PluralRules, PluralRuleType, PluralCategory};
use icu_provider::inv::InvariantDataProvider;

let dp = InvariantDataProvider;

let pr = PluralRules::try_new(locale!("en"), &dp, PluralRuleType::Cardinal)
    .expect("Failed to construct a PluralRules struct.");

assert_eq!(pr.select(5_usize), PluralCategory::Other);

Implementations

impl PluralRules

pub fn try_new<T: Into<Locale>, D>(
locale: T,
data_provider: &D,
rule_type: PluralRuleType
) -> Result<Self, PluralRulesError> where
D: ResourceProvider<CardinalV1Marker> + ResourceProvider<OrdinalV1Marker> + ?Sized,

Constructs a new PluralRules for a given locale, type and data provider.

This constructor will fail if the Data Provider does not have the data.

Examples

use icu::locid::locale;
use icu::plurals::{PluralRules, PluralRuleType};
use icu_provider::inv::InvariantDataProvider;

let lid = locale!("en");

let dp = InvariantDataProvider;

let _ = PluralRules::try_new(lid, &dp, PluralRuleType::Cardinal);

pub fn try_new_cardinal<T: Into<Locale>, D>(
locale: T,
data_provider: &D
) -> Result<Self, PluralRulesError> where
D: ResourceProvider<CardinalV1Marker> + ?Sized,

Constructs a new PluralRules for a given locale for cardinal numbers.

Cardinal plural forms express quantities of units such as time, currency or distance, used in conjunction with a number expressed in decimal digits (i.e. “2”, not “two”).

For example, English has two forms for cardinals:

One: 1 day

Other: 0 days, 2 days, 10 days, 0.3 days

Examples

use icu::locid::locale;
use icu::plurals::{PluralRules, PluralCategory};

let dp = icu_testdata::get_provider();

let rules = PluralRules::try_new_cardinal(locale!("ru"), &dp)
    .expect("Data should be present");

assert_eq!(rules.select(2_usize), PluralCategory::Few);

pub fn try_new_ordinal<T: Into<Locale>, D>(
locale: T,
data_provider: &D
) -> Result<Self, PluralRulesError> where
D: ResourceProvider<OrdinalV1Marker> + ?Sized,

Constructs a new PluralRules for a given locale for ordinal numbers.

Ordinal plural forms denote the order of items in a set and are always integers.

For example, English has four forms for ordinals:

One: 1st floor, 21st floor, 101st floor

Two: 2nd floor, 22nd floor, 102nd floor

Few: 3rd floor, 23rd floor, 103rd floor

Other: 4th floor, 11th floor, 96th floor

Examples

use icu::locid::locale;
use icu::plurals::{PluralRules, PluralCategory};

let dp = icu_testdata::get_provider();

let rules = PluralRules::try_new_ordinal(locale!("ru"), &dp)
    .expect("Data should be present");

assert_eq!(rules.select(2_usize), PluralCategory::Other);

pub fn select<I: Into<PluralOperands>>(&self, input: I) -> PluralCategory

Returns the Plural Category appropriate for the given number.

Examples

use icu::locid::locale;
use icu::plurals::{PluralRules, PluralRuleType, PluralCategory};
use icu_provider::inv::InvariantDataProvider;

let dp = InvariantDataProvider;

let pr = PluralRules::try_new(locale!("en"), &dp, PluralRuleType::Cardinal)
    .expect("Failed to construct a PluralRules struct.");

match pr.select(1_usize) {
    PluralCategory::One => "One item",
    PluralCategory::Other => "Many items",
    _ => { unreachable!(); }
};

The method accepts any input that can be calculated into Plural Operands. All unsigned primitive number types can infallibly be converted so they can be used as an input.

For signed numbers and strings, Plural Operands implement TryFrom and FromStr, which should be used before passing the result to select().

Examples

use std::convert::TryFrom;
use icu::locid::locale;
use icu::plurals::{PluralRules, PluralRuleType};
use icu::plurals::{PluralCategory, PluralOperands};
use icu_provider::inv::InvariantDataProvider;

let operands = PluralOperands::try_from(-5)
    .expect("Failed to parse to operands.");
let operands2: PluralOperands = "5.10".parse()
    .expect("Failed to parse to operands.");

assert_eq!(pr.select(operands), PluralCategory::Other);
assert_eq!(pr.select(operands2), PluralCategory::Other);

pub fn categories(&self) -> impl Iterator<Item = PluralCategory> + '_

Returns all Plural Categories appropriate for a PluralRules object based on the LanguageIdentifier and PluralRuleType.

The Plural Categories are returned in UTS 35 sorted order.

The category PluralCategory::Other is always included.

Examples

use icu::locid::locale;
use icu::plurals::{PluralRules, PluralRuleType, PluralCategory};
use icu_provider::inv::InvariantDataProvider;

let dp = icu_testdata::get_provider();

let pr = PluralRules::try_new(locale!("fr"), &dp, PluralRuleType::Cardinal)
    .expect("Failed to construct a PluralRules struct.");

let mut categories = pr.categories();
assert_eq!(categories.next(), Some(PluralCategory::One));
assert_eq!(categories.next(), Some(PluralCategory::Many));
assert_eq!(categories.next(), Some(PluralCategory::Other));
assert_eq!(categories.next(), None);